requirements.txt
pillow
opencv-python
numpy
torch
torchvision
pyyaml
kornia
tqdm
matplotlib
requests
gdown